bug: implemented background for certain feature tests, fixed pending test
Faculty-related feature tests were not making use of the background section, causing a significant amount of clutter. This has been fixed, and certain step definitions were changed to be less redundant. Lastly, a test was unnecessarily marked as pending due to an implementation detail that meant two separate cases would arise depending on filtering. This test has had conditions added to it to alleviate this.